# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Message": {
        "additionalProperties": false,
        "description": "One outbound business communication to a person.\n\n`extra=\"forbid\"`: this model is the single source of truth for the comms-history\nshape \u2014 a field that isn't declared here doesn't exist.",
        "properties": {
          "body": {
            "default": "",
            "title": "Body",
            "type": "string"
          },
          "channel": {
            "$ref": "#/components/schemas/MessageChannel",
            "default": "email"
          },
          "contact_id": {
            "default": "",
            "title": "Contact Id",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "listing_key": {
            "default": "",
            "title": "Listing Key",
            "type": "string"
          },
          "provider_message_id": {
            "default": "",
            "title": "Provider Message Id",
            "type": "string"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "status": {
            "$ref": "#/components/schemas/MessageStatus",
            "default": "drafted"
          },
          "subject": {
            "default": "",
            "title": "Subject",
            "type": "string"
          },
          "template_ref": {
            "default": "",
            "title": "Template Ref",
            "type": "string"
          },
          "tenant_id": {
            "default": "",
            "title": "Tenant Id",
            "type": "string"
          },
          "transaction_id": {
            "default": "",
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "recipient"
        ],
        "title": "Message",
        "type": "object"
      },
      "MessageChannel": {
        "enum": [
          "email",
          "sms"
        ],
        "title": "MessageChannel",
        "type": "string"
      },
      "MessageStatus": {
        "description": "Lifecycle of an outbound message.\n\nDRAFTED \u2192 SENT | FAILED is the direct-send path (BOP-015). Workflow-drafted\ncomms (BOP-019) insert a human gate between draft and send:\nPENDING_APPROVAL \u2192 SENT | FAILED | REJECTED \u2014 no drafted text leaves the\nboundary without a human decision. DELIVERED is set only by a provider\ndelivery-status webhook (BOP-018): SENT means the provider accepted the\nmessage; DELIVERED means the provider confirmed the handset got it. A\ndelivery callback can also move SENT \u2192 FAILED.",
        "enum": [
          "drafted",
          "pending_approval",
          "sent",
          "delivered",
          "failed",
          "rejected"
        ],
        "title": "MessageStatus",
        "type": "string"
      },
...
      "SendMessageRequest": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/MessageChannel",
            "default": "email"
          },
          "contact_id": {
            "default": "",
            "title": "Contact Id",
            "type": "string"
          },
          "listing_key": {
            "default": "",
            "title": "Listing Key",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Params",
            "type": "object"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "request_id": {
            "default": "",
            "title": "Request Id",
            "type": "string"
          },
          "template": {
            "title": "Template",
            "type": "string"
          },
          "transaction_id": {
            "default": "",
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "recipient",
          "template"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "brokerops api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/messages": {
      "get": {
        "description": "The outbound comms history, newest first \u2014 read-open like the audit trail.\n\n`transaction_id` scopes the history to one deal for the transaction hub\n(BOP-027); both filters compose (AND) when supplied together. The response is\ncaller-role filtered (BOP-040): the freeform body/subject and the recipient\n(contact PII) are redacted for viewers so a viewer-open hub never receives\nmessage content over the wire.",
        "operationId": "list_messages_messages_get",
        "parameters": [
          {
            "in": "query",
            "name": "contact_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contact Id"
            }
          },
          {
            "in": "query",
            "name": "transaction_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Transaction Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  },
                  "title": "Response List Messages Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Messages",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/send": {
      "post": {
        "operationId": "send_message_messages_send_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/{message_id}": {
      "get": {
        "operationId": "get_message_messages__message_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/messages/{message_id}/retry": {
      "post": {
        "description": "Re-drive a FAILED send to completion (BOP-037).\n\nThe gap this closes: an approved gate whose provider send failed leaves the\napproval APPROVED (the decide route 409s forever) and the row FAILED \u2014 with\nno surface able to reach `send_approved`'s FAILED\u2192SENT path. Strictly\nFAILED-only, so it can never bypass a pending human gate or re-send a\nterminal row; the send runs through the same seam (audited + deduped) under\na fresh run id, linked back to the original approval when one exists.",
        "operationId": "retry_failed_message_messages__message_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Failed Message",
        "tags": [
          "messages"
        ]
      }
    },
...
  }
}